--- CustomSabers/Data/WhackerTrail.cs
@@ -1,14 +1,0 @@
-﻿using UnityEngine;
-
-namespace CustomSabersLite.Data
-{
-    internal struct WhackerTrail
-    {
-        public int trailId;
-        public ColorType colorType;
-        public Color trailColor;
-        public Color multiplierColor;
-        public int length;
-        public float whiteStep;
-    }
-}
